fix `target` attribute for log_impl!
The __private_api_log takes separate argments for the "target" and the
"module path" but the second match case for the log_impl macros was
passing the module path for both. This means a log call like:

   info!(target: "my_custom_target", "message", {kv_pairs: values});

would discard the custom target and be sent to the target with the
module's name.

The first case of this macro without any kv pairs is already correctly
using the target.